la modification dans les roles des users

+++ pepiniere/routes/api.php
@@ -36,17 +36,19 @@ use Illuminate\Support\Facades\Route;
     Route::post('/login', [AuthController::class, 'login']);
     Route::post('/store', [CategorieController::class, 'store']);
     Route::post('/plantes', [PlanteController::class, 'storePlante']);
-    Route::get('/categories',[CategorieController::class,'afficherCategorie']);
-    Route::get('/getplantes/{categorie_id}', [PlanteController::class, 'afficherPlantes']);
-    Route::put('/modifierCategorie/{id}', [CategorieController::class, 'modifierCategorie']);
-    Route::put('/modifierPlante/{id}', [PlanteController::class, 'modifierPlante']);
-    Route::delete('/supprimerCategorie/{id}', [CategorieController::class, 'supprimerCategorie']);
-    Route::delete('/supprimerPlante/{id}', [PlanteController::class, 'supprimerPlante']);
-    Route::get('/afficherPlanteDetailles/{slug}',[PlanteController::class,'afficherPlanteDetailles']);
+     // jai utilsé ici un middelware pour verifier si il est authentifié //
+    Route::middleware(['check.auth'])->get('/categories', [CategorieController::class, 'afficherCategorie']);
 
-Route::middleware('auth:api')->group(function () {
+    Route::get('/getplantes/{categorie_id}', [PlanteController::class, 'afficherPlantes']);
+    Route::middleware(['check.role:admin'])->put('/modifierCategorie/{id}', [CategorieController::class, 'modifierCategorie']);
+    Route::middleware(['check.role:admin'])->put('/modifierPlante/{id}', [PlanteController::class, 'modifierPlante']);
+    Route::middleware(['check.role:admin'])->delete('/supprimerCategorie/{id}', [CategorieController::class, 'supprimerCategorie']);
+    Route::middleware(['check.role:admin'])->delete('/supprimerPlante/{id}', [PlanteController::class, 'supprimerPlante']);
+    Route::get('/afficherPlanteDetailles/{slug}',[PlanteController::class,'afficherPlanteDetailles']);
+    Route::post('/AjouterCommande/{slug}',[CommandeController::class,'AjouterCommande']);
+
 
     Route::post('/logout', [AuthController::class, 'logout']);
     Route::get('/me', [AuthController::class, 'me']);
     
-    });
+
